crypto-ha-bot-v8/bot/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
pymysql==1.1.1
python-dotenv==1.0.1
pydantic==2.7.4
requests==2.32.3
binance-connector==3.3.0
yfinance==0.2.40
pandas==2.2.2
statsmodels==0.14.2
ccxt==4.5.1
